git: all_commits_...: make `git describe` optional
`git log --format=format:%(describe)` can be very slow because git CLI
spawns a subprocess for every commit. It's by far the slowest part of
this function, so let's make it optional.

Also, change the order of the function's return values, so that the now
optional `commit_desc` part comes last.